Implement Access Control

+++ app/Http/Controllers/PostsController.php
@@ -11,6 +11,15 @@ use DB;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -34,6 +43,7 @@ class PostsController extends Controller
      */
     public function create()
     {
+        //return Post::all;
         return view('posts.create');
     }
 
@@ -81,6 +91,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -116,6 +132,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        //check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Sorry you are not authorized for that action');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
